+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.2] - Unreleased
 
+### Added
+
+- Add a `reissue:branch` task to create a new branch for the next version
+- Add a `reissue:push` task to push the new branch to the remote repository
+
 ### Fixed
 
 - Require the 'date' library in tests to ensure the Date constant is present